webui: dislay server suffixes in server search page
There was a change where suffixes in server are not longer returned as DNs
but rather a cn of related topology suffix. I.e. they share "memberof" logic.
This caused that search page doesn't get the data because it uses
"no_member: true" option by default.

This patch overrides the behavior because it is OK for server search page
to fetch also member data - it is not so costly as e.g. in users.

https://fedorahosted.org/freeipa/ticket/5609

Reviewed-By: Martin Babinsky <mbabinsk@redhat.com>