Fix ipa-replica-install with RHEL 6.8 master
ipa-replica-prepare creates a gpg file containing realm_info/cacert.p12 with
the certificates.
When run on a RHEL 6.8 instance, cacert.p12 contains twice the same cert
(for caSigningCert cert-pki-ca), once with the nickname and once without.

ipa-replica-install passes this file to pkispawn and makes pkispawn fail.
The fix exports the pkcs12 file into a temp nsddb then re-creates a pkcs12
file from the nssdb (this process removes the duplicate cert).

https://fedorahosted.org/freeipa/ticket/6310