Mark all phony Makefile targets as .PHONY
https://www.gnu.org/software/make/manual/make.html#Phony-Targets

A phony target is one that is not really the name of a file; rather it
is just a name for a recipe to be executed when you make an explicit
request. There are two reasons to use a phony target: to avoid a
conflict with a file of the same name, and to improve performance.

Signed-off-by: Christian Heimes <cheimes@redhat.com>